realizando o crud completo

+++ src/main/java/org/academify/model/repository/AlunoRepository.java
@@ -3,5 +3,5 @@ package org.academify.model.repository;
 import org.academify.model.entity.Aluno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlunoRepository extends JpaRepository<Aluno, Integer> {
+public interface AlunoRepository extends JpaRepository<Aluno, Long> {
 }
